pyproject.toml
[project]
name = "diashow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"

dependencies = [
    "pillow>=11.3.0",
    "watchdog>=6.0.0",
]

[tool.ruff]
# Set the maximum line length to 120
line-length = 120

# Enable specific rule sets
select = [ "ALL" ]

# Ignore specific rules
ignore = [
    "E501",  # line too long (handled by line-length)
    "LOG015", # root-logger-call
    "D212", 
    "D205",
    "D107"
]

# Exclude directories
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

[tool.ruff.format]
# Use double quotes for strings
quote-style = "double"

# Indent with spaces, rather than tabs
indent-style = "space"
